+++ code/src/layouts/header/components/header-toolbar/HeaderToolbar.test.tsx
@@ -62,4 +62,5 @@ describe("HeaderToolbar", () => {
       expect(mockDispatch).toHaveBeenCalledWith(logout());
     });
   });
+  
 });
